SimpleStorage/scripts/deploy.py
import os
from brownie import accounts, config, SimpleStorage, network

def get_account():
    if network.show_activate() == "development":
        return accounts[0]
    else:
        return accounts.add(config["wallets"]["from_key"])

def deploy_simple_storage():
    # account = accounts.load("naiyoma-account")
    # account = accounts.add(os.getenv("PRIVATE_KEY"))
    # account = accounts.add(config["wallets"]["from_key"])
    account = get_account()
    simple_storage = SimpleStorage.deploy({"from": account})
    stored_value = simple_storage.retrieve()
    transaction = simple_storage.store(15, {"from": account})
    transaction.wait(1)
    stored_value = simple_storage.retrieve()



def main():
    deploy_simple_storage()
