requirements.txt
pkg-resources==0.0.0
py-solc==3.2.0
py-solc-x==1.1.1
requests==2.26.0
semantic-version==2.8.5
urllib3==1.26.7
web3==5.24.0
python-dotenv==0.19.1
vvm==0.1.0
vyper==0.2.16
py-solc==3.2.0
py-solc-ast==1.2.9
py-solc-x==1.1.0
eth-abi==2.1.1
eth-account==0.5.5
eth-brownie==1.16.4
eth-event==1.2.3
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
